Update tKey example
